update UI for main profile page
