+++ src/app/profile/[userId]/_components/Profile/PhotoAlbum.tsx
@@ -4,10 +4,20 @@ import { ImageProp } from '@/server/routers/Images/index';
 
 export default function PhotoAlbum({ images }: { images: Array<ImageProp> }) {
   if (images.length === 0) {
-    return <div>No images under Photo Album</div>;
+    return (
+      <div
+        className="flex min-h-80 items-center justify-center rounded-b-xl rounded-tr-xl"
+        style={{ backgroundColor: '#F7F5EF' }}
+      >
+        <p>No images under Photo Album</p>
+      </div>
+    );
   }
   return (
-    <main className="flex flex-col gap-4">
+    <main
+      className="flex flex-col gap-4 rounded-b-xl rounded-tr-xl p-10"
+      style={{ backgroundColor: '#F7F5EF' }}
+    >
       <div className="grid grid-cols-1 gap-10 md:grid-cols-2 lg:grid-cols-4">
         {images.map((image) => (
           <div key={image.id} className="relative flex h-72 gap-4 p-4">
