+++ src/app/profile/[userId]/_components/Profile/Projects.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useEffect } from 'react';
 import { trpc } from '@/lib/trpc/client';
-import { Button } from '@/components/ui/button';
 import { useState } from 'react';
 import PhotoAlbum from './PhotoAlbum';
 
@@ -31,18 +30,17 @@ export default function Projects({ clerkId }: { clerkId: string }) {
   }
 
   return (
-    <main className="mt-10 flex flex-col gap-4">
-      <h2 className="font-mono text-xl font-bold">Projects</h2>
-      <div className="flex flex-row gap-x-4">
+    <main className="mt-20 flex flex-col">
+      <div className="flex flex-row">
         {photoAlbums &&
           photoAlbums.map((photoAlbum, index) => (
-            <Button
+            <button
               key={photoAlbum.id}
-              className="rounded-full border border-gray-400 bg-profile_button_bg text-xs text-black hover:bg-primary_blue hover:text-white active:bg-primary_blue active:text-white"
+              className="rounded-t-2xl px-10 py-2.5 text-sm text-black shadow-md hover:contrast-75"
               style={
                 selectedPhotoAlbum?.photoAlbumId === photoAlbum.id
-                  ? { backgroundColor: '#013C5A', color: 'white' }
-                  : {}
+                  ? { backgroundColor: '#F7F5EF' }
+                  : { backgroundColor: '#E6E8E9' }
               }
               aria-label={`Photo Album ${photoAlbum.photoAlbumName}`}
               onClick={() =>
@@ -52,8 +50,8 @@ export default function Projects({ clerkId }: { clerkId: string }) {
                 })
               }
             >
-              {photoAlbum.photoAlbumName} ({photoAlbum.Images.length})
-            </Button>
+              {photoAlbum.photoAlbumName}
+            </button>
           ))}
       </div>
       {selectedPhotoAlbum && (
