+++ src/app/profile/[userId]/page.tsx
@@ -2,7 +2,7 @@ import Profile from '@/app/profile/[userId]/_components/Profile/Profile';
 
 export default function Page() {
   return (
-    <main className="pb-auto w-full max-w-7xl px-4 pb-4 sm:px-10 sm:pb-10 md:px-12 md:pb-12 2xl:px-0">
+    <main className="pb-auto w-full max-w-7xl pb-4 sm:pb-10 md:pb-12 ">
       <Profile />
     </main>
   );
